src/jobanalytics/requirements.txt
Babel==2.5.3
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
coverage==4.5.1
cycler==0.10.0
decorator==4.1.2
dj-static==0.0.6
Django==1.11.6
django-phonenumber-field==2.0.0
django-tinymce==2.7.0
djangorestframework==3.8.2
docx2txt==0.7
fuzzywuzzy==0.16.0
graphviz==0.8.1
idna==2.6
ijson==2.3
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.3.0
kmodes==0.8
lxml==4.1.1
matplotlib==2.1.0
nbformat==4.4.0
nltk==3.2.5
numpy==1.13.0
pandas==0.20.3
phonenumberslite==8.8.11
plotly==2.1.0
probscale==0.2.3
psycopg2==2.7.3.2
pyparsing==2.2.0
PyPDF2==1.26.0
PyQt5==5.10.1
python-dateutil==2.6.1
python-docx==0.8.6
python-Levenshtein==0.12.0
pytz==2017.2
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
sip==4.19.8
six==1.11.0
sklearn==0.0
static3==0.7.0
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
